requirements.txt
streamlit==1.49.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
python-dotenv==1.0.0
pypdf==4.0.0
docarray==0.41.0
sentence-transformers==2.2.2
torch==2.0.0
numpy==1.24.0